+++ src/mon-util.h
@@ -35,6 +35,7 @@ bool monster_carry(struct chunk *c, struct monster *mon, struct object *obj);
 void monster_swap(struct loc grid1, struct loc grid2);
 void monster_wake(struct monster *mon, bool notify, int aware_chance);
 void monster_set_allegiance(struct monster *mon, enum monster_allegiance side);
+void monster_make_hostile(struct monster *mon, const char *what);
 bool monster_can_see(struct chunk *c, struct monster *mon, struct loc grid);
 void become_aware(struct chunk *c, struct monster *m);
 void update_smart_learn(struct monster *mon, struct player *p, int flag,